requirements.txt
openai==0.27.10
pandas==1.5.3
pymongo==4.5.0
python-dotenv==1.0.1
Requests==2.31.0
streamlit==1.31.0
